Fixing SQL

+++ project/sql/006_create_table_transactions.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS `Transactions` (
     `id`                    INT NOT NULL AUTO_INCREMENT,
     `account_source`        INT NOT NULL,
-    `account_destination`   DECIMAL(12, 2) NOT NULL,
+    `account_destination`   INT NOT NULL,
     `balance_change`        DECIMAL(12, 2) NOT NULL,
     `transaction_type`      VARCHAR(12),
     `memo`                  VARCHAR(150) DEFAULT '',
